Добавил восстановление после переподключения
